+++ src/module/vue/sfmove-sheet.vue
@@ -9,13 +9,13 @@
       <div class="flexcol nogrow" style="white-space: nowrap; max-width: 33%">
         <!-- These are always here -->
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="FullText"
           property="Text"
         />
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Trigger"
           property="Trigger.Text"
@@ -30,7 +30,7 @@
         <sfmove-tab
           v-for="option in triggerOptions"
           :key="option.key"
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           :title="option.title"
           :property="option.property"
           @click="switchContent($event, option.actionPropKey)"
@@ -41,31 +41,31 @@
         <hr class="nogrow" />
         <h4 class="nogrow">Outcomes</h4>
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Strong_hit"
           property="Outcomes.Strong Hit.Text"
         />
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Strong_hit_match"
           property="Outcomes.Strong Hit.With a Match.Text"
         />
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Weak_hit"
           property="Outcomes.Weak Hit.Text"
         />
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Miss"
           property="Outcomes.Miss.Text"
         />
         <sfmove-tab
-          :currentProperty="currentProperty"
+          :currentProperty="state.currentProperty"
           @click="switchContent"
           titleKey="Miss_match"
           property="Outcomes.Miss.With a Match.Text"
@@ -74,7 +74,7 @@
 
       <!-- Editor on right -->
       <div class="flexcol">
-        <div class="flexcol nogrow" v-if="currentRollType">
+        <div class="flexcol nogrow" v-if="state.currentRollType">
           <div class="flexrow">
             <div class="flexcol">
               <label
@@ -86,7 +86,7 @@
                   type="radio"
                   name="actiontype"
                   :value="x"
-                  v-model="currentRollType"
+                  v-model="state.currentRollType"
                   @change="saveActionProps"
                 />
                 {{ x.split(' ')[0] }}
@@ -102,16 +102,16 @@
                   type="radio"
                   name="rollType"
                   :value="x"
-                  v-model="currentMethod"
+                  v-model="state.currentMethod"
                   @change="saveActionProps"
                 />
                 {{ x }}
               </label>
             </div>
           </div>
-          <input v-model="currentStatText" @blur="saveActionProps" />
+          <input v-model="state.currentStatText" @blur="saveActionProps" />
         </div>
-        <textarea v-model="currentContent" @blur="saveText" />
+        <textarea v-model="state.currentContent" @blur="saveText" />
       </div>
     </div>
   </div>
@@ -127,99 +127,104 @@
 }
 </style>
 
-<script>
+<script setup lang="ts">
+import { provide, computed, reactive, inject } from 'vue'
 import { get, set } from 'lodash'
+import { $ItemKey } from './provisions'
 import DocumentName from './components/document-name.vue'
 import SfmoveTab from './components/sfmove-tab.vue'
 import BtnFaicon from './components/buttons/btn-faicon.vue'
 
-export default {
-  props: {
-    item: Object,
-  },
-  data() {
+const props = defineProps<{
+  item: any
+}>()
+provide(
+  'item',
+  computed(() => props.item)
+)
+const $item = inject($ItemKey)
+
+const state = reactive({
+  currentProperty: 'Text',
+  currentActionPropKey: undefined,
+  currentContent: props.item.data.Text,
+  currentRollType: undefined,
+  currentMethod: undefined,
+  currentStatText: undefined,
+})
+
+const triggerOptions = computed(() => {
+  const itemTriggerOptions = props.item.data.Trigger?.Options || []
+  return itemTriggerOptions.map((x, i) => {
+    const title = x['Action roll']
+      ? `Roll +${x['Action roll'].Stat}`
+      : `${i + 1}`
     return {
-      currentProperty: 'Text',
-      currentActionPropKey: undefined,
-      currentContent: this.item.data.Text,
-      currentRollType: undefined,
-      currentMethod: undefined,
-      currentStatText: undefined,
+      key: `option${i}`,
+      title,
+      actionPropKey: `Trigger.Options[${i}]`,
+      property: `Trigger.Options[${i}].Text`,
     }
-  },
-  computed: {
-    triggerOptions() {
-      const itemTriggerOptions = this.item.data.Trigger?.Options || []
-      return itemTriggerOptions.map((x, i) => {
-        const title = x['Action roll']
-          ? `Roll +${x['Action roll'].Stat}`
-          : `${i + 1}`
-        return {
-          key: `option${i}`,
-          title,
-          actionPropKey: `Trigger.Options[${i}]`,
-          property: `Trigger.Options[${i}].Text`,
-        }
-      })
-    },
-  },
-  methods: {
-    switchContent(prop, actionPropKey = undefined) {
-      this.currentProperty = prop
-      this.currentContent = get(this.item.data, prop)
-      this.currentActionPropKey = actionPropKey
-      // {
-      //   Method: 'Any',
-      //   'Roll type': 'Action roll',
-      //   Text: 'Receive treatment from someone (not an ally)',
-      //   Using: ['Iron'],
-      //   dfid: 'Starforged/Moves/Recover/Heal/Trigger/Options/1',
-      // }
-      const ap = actionPropKey && get(this.item.data, actionPropKey)
-      this.currentRollType = ap?.['Roll type']
-      this.currentMethod = ap?.Method
-      this.currentStatText = ap?.Using?.join?.(',') ?? ''
-    },
-    addTrigger() {
-      let { Options } = this.item.data.Trigger
-      Options ||= []
-      Options.push({
-        Text: '',
-        Method: 'Any',
-        'Roll type': 'Action roll',
-        Using: 'Iron',
-      })
-      this.$item.update({ data: { Trigger: { Options } } })
-    },
-    removeTrigger(option) {
-      console.log(option)
-      const idx = this.triggerOptions.findIndex((x) => x.key === option.key)
-      let { Options } = this.item.data.Trigger
-      Options ||= []
-      Options.splice(idx, 1)
-      this.$item.update({ data: { Trigger: { Options } } })
-      this.switchContent('Text')
-    },
-    saveActionProps() {
-      const opt = get(this.item.data, this.currentActionPropKey)
-      opt.Method = this.currentMethod
-      opt['Roll type'] = this.currentRollType
-      opt.Using = this.currentStatText.split(',').map((x) => x.trim())
-      set(this.item.data, this.currentActionPropKey, opt)
-      this.$item.update({ data: this.item.data })
-    },
-    saveText() {
-      if (this.currentProperty.includes('Options')) {
-        set(this.item.data, this.currentProperty, this.currentContent)
-        const { Options } = this.item.data.Trigger
-        this.$item.update({ data: { Trigger: { Options } } })
-      } else {
-        this.$item.update({
-          data: { [this.currentProperty]: this.currentContent },
-        })
-      }
-    },
-  },
-  components: { DocumentName, SfmoveTab, BtnFaicon },
+  })
+})
+
+function switchContent(prop, actionPropKey = undefined) {
+  state.currentProperty = prop
+  state.currentContent = get(props.item.data, prop)
+  state.currentActionPropKey = actionPropKey
+  // {
+  //   Method: 'Any',
+  //   'Roll type': 'Action roll',
+  //   Text: 'Receive treatment from someone (not an ally)',
+  //   Using: ['Iron'],
+  //   dfid: 'Starforged/Moves/Recover/Heal/Trigger/Options/1',
+  // }
+  const ap = actionPropKey && get(props.item.data, actionPropKey)
+  state.currentRollType = ap?.['Roll type']
+  state.currentMethod = ap?.Method
+  state.currentStatText = ap?.Using?.join?.(',') ?? ''
+}
+
+function addTrigger() {
+  let { Options } = props.item.data.Trigger
+  Options ||= []
+  Options.push({
+    Text: '',
+    Method: 'Any',
+    'Roll type': 'Action roll',
+    Using: 'Iron',
+  })
+  $item?.update({ data: { Trigger: { Options } } })
+}
+
+function removeTrigger(option) {
+  console.log(option)
+  const idx = state.triggerOptions.findIndex((x) => x.key === option.key)
+  let { Options } = props.item.data.Trigger
+  Options ||= []
+  Options.splice(idx, 1)
+  $item?.update({ data: { Trigger: { Options } } })
+  switchContent('Text')
+}
+
+function saveActionProps() {
+  const opt = get(props.item.data, state.currentActionPropKey)
+  opt.Method = state.currentMethod
+  opt['Roll type'] = state.currentRollType
+  opt.Using = state.currentStatText.split(',').map((x) => x.trim())
+  set(props.item.data, state.currentActionPropKey, opt)
+  $item?.update({ data: props.item.data })
+}
+
+function saveText() {
+  if (state.currentProperty.includes('Options')) {
+    set(props.item.data, state.currentProperty, state.currentContent)
+    const { Options } = props.item.data.Trigger
+    $item?.update({ data: { Trigger: { Options } } })
+  } else {
+    $item?.update({
+      data: { [state.currentProperty]: state.currentContent },
+    })
+  }
 }
 </script>